Refactor AddBoxForm to AdminModal
AddBox does not need from and its going to be replaced with admin only functionality. Was renamed and moved into a new folder for modals.
